Added.

+++ Bitwise_Oparators/Data_type_modifiers/short.cpp
@@ -4,6 +4,6 @@ int main(){
 ios_base::sync_with_stdio(false);
 cin.tie(nullptr);
     cout<<sizeof(int)<<endl;
-    cout<<sizeof(long long int)<<endl;
+    cout<<sizeof(short int)<<endl;
     return 0;
 }
